requirements.txt
streamlit
google-generativeai
chromadb
requests
streamlit
google-generativeai
chromadb
requests
python-dotenv
pymongo
